- vector and supporting functions moved to new core folder vector. - deleted sum.hpp
