+++ include/NeoN/core/database/fieldCollection.hpp
@@ -8,7 +8,7 @@
 
 #include "NeoN/core/demangle.hpp"
 #include "NeoN/core/error.hpp"
-#include "NeoN/core/vector.hpp"
+#include "NeoN/core/vector/vector.hpp"
 #include "NeoN/fields/field.hpp"
 #include "NeoN/core/database/database.hpp"
 #include "NeoN/core/database/collection.hpp"
