+++ include/NeoN/core/vector/vectorTypeDefs.hpp
@@ -4,7 +4,7 @@
 
 #include "NeoN/core/primitives/scalar.hpp"
 #include "NeoN/core/primitives/vec3.hpp"
-#include "NeoN/core/vector.hpp"
+#include "NeoN/core/vector/vector.hpp"
 
 
 namespace NeoN
