+++ include/NeoN/dsl/coeff.hpp
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "NeoN/core/primitives/scalar.hpp"
-#include "NeoN/core/vector.hpp"
+#include "NeoN/core/vector/vector.hpp"
 
 namespace NeoN::dsl
 {
