+++ include/NeoN/fields/boundaryData.hpp
@@ -8,7 +8,7 @@
 #include "NeoN/core/primitives/label.hpp"
 #include "NeoN/core/primitives/scalar.hpp"
 #include "NeoN/core/executor/executor.hpp"
-#include "NeoN/core/vector.hpp"
+#include "NeoN/core/vector/vector.hpp"
 
 #include <vector>
 #include <utility>
