+++ include/NeoN/linearAlgebra/linearSystem.hpp
@@ -2,7 +2,7 @@
 // SPDX-FileCopyrightText: 2023 NeoN authors
 #pragma once
 
-#include "NeoN/core/vector.hpp"
+#include "NeoN/core/vector/vector/vector.hpp"
 #include "NeoN/linearAlgebra/CSRMatrix.hpp"
 
 #include <string>
