+++ src/core/vector/vector.cpp
@@ -3,7 +3,7 @@
 
 #include "NeoN/core/primitives/vec3.hpp"
 #include "NeoN/core/primitives/scalar.hpp"
-#include "NeoN/core/vector.hpp"
+#include "NeoN/core/vector/vector.hpp"
 
 #include <variant>
 #include <vector>
